Using the new logical 50pin connector.
The new "logical" 50pin lcd connector is bit clearer layed out. (based
on adafruit symbol) also we added a few jumpers in case we use the same
pinout lcd as adafruit and not the one from buydisplay. The buydisplay
capacitive touch display uses a separate connector for the captouch
sensor data lines.