Added USB V+ pogopin pad. Finalized battery management circuit.
The battery charging and management circuit will be using the USB V+
coming from an SMD pad tapped off by a pogopin. The 1Bitsy will require
a modification removing the diode connecting USB V+ to +5V net. The
powerpath output of the charge and management chip is then connected to
the +5V rail.