Keys are now connected through shift registers.
To save some pins I decided to connect the buttons through a shift
register in the style of NES/SNES controllers. I could not find an SPI
interface on the STM32 that would offer MISO, NSS and SCK without
colliding with some other feature of the 1UP.

The final compromise I settled on is to connect SCK to SPI1MOSI and DO
to SPI1MISO. We can either just bitbang the shift register shifting or
send out alternating 01 patterns and read in 4 bits from the shift
registers for every 8 bits shifted out. There might be some other more
efficient hackery that could be done by starting a timer 50% duty cycle
signal out to clock the shift registers and have the MISO sample the
bits at the same time. Some interesting hacks pending. :D