Added volume up/down buttons.
I wanted to use a thumbwheel encoder, but just like thumbwheel
potentiometers they don't seem te be in production any more. (philips
used to be the main manufacturer of those but they seem to have
discontinued them) Decided to go with up down buttons instead. This
helps us with not having to use up more of the 1Bitsy GPIO which is not
bad.