Placed mechanical tft footprint. Some mechanical rearangement.
Using the new mechanical tft drawing, I have placed the position of the
TFT on the board. Using those references the 50pin and 6pin connectors
found their positions. Also had to move the micro SD card slot as it was
colliding with the 6pin capacitive touch socket. Added the cutout for
the flatflex cable. Moved 1bitsy so that it does not interfere with the
TFT. And updated PKL.