Distributed passives to their respective active components & schematic fixes.
Moved all the passives to the area where their active component circuits
reside at the moment. The position of the subcircuits still can change
and move.

Also fixed the backlight mosfet circuit. (the sot23 pads were not
connecting to the net because of a different pad naming scheme) Also
found that the VCC of the lcd logic was not connected to power. DUH