Final version & Added backplate mounting holes, moved heart to mask layer.
This is the final prototype version of the board. Submitted to OSHPark
for production.

A great suggestion from twitter, is to add a back plate to the board, to
make that possible I have added a few more m3 mounting holes.

When exporting gerber files I have found out that the pixel heart is not
showing up, it turns out polygons do not work on silkscreen for some
reason. (I bet it is some KiCad bug) Moved it back to solder stop layer
again. (I have no problems with golden hearts! :D )