Changed lcd orientation to match hardware.
We use the LCD in landscape mode, also I somehowe ended up mounting it
up side down? o_O ... so we needed to tell the lcd to swap axis, and
reverse drawing order on both axis. As a result we have the expected top
to bottom and left to right pixel order for the data we are sending to
the display. :D