Fixed battery charge/management circuit.
The SYSOFF pin is meant for disconnecting the battery from the system,
but not for poweroff operation but for battery monitoring purposes. For
example fuel gauge chip applications. The previous circuit wos
preventing the 1UP from charging the battery while the system was
powered off.

While we were at it I also updated the schematic to KiCad V5.0.